+++ src/Command/Install/CreatePluginsLinksCommand.php
@@ -49,7 +49,7 @@ class CreatePluginsLinksCommand extends Command
         foreach (Plugin::loaded() as $plugin) {
             $srcPath = Plugin::path($plugin) . 'webroot';
             if (!is_dir($srcPath)) {
-                $io->verbose(sprintf('Skipping plugin %s. It does not have webroot folder.', $plugin), 1);
+                $io->verbose(__d('me_tools', 'Skipping plugin `{0}`. It does not have webroot folder', $plugin), 1);
                 continue;
             }
 
@@ -59,17 +59,8 @@ class CreatePluginsLinksCommand extends Command
 
         foreach ($plugins as $plugin => $config) {
             $io->verbose('For plugin: ' . $plugin);
-
-            $dest = $config['destDir'] . $config['link'];
-            if (file_exists($dest)) {
-                $io->verbose('Link `' . rtr($dest) . '` already exists', 1);
-                continue;
-            }
-
-            $this->createLink($io, $config['srcPath'], $dest);
+            $this->createLink($io, $config['srcPath'], $config['destDir'] . $config['link']);
         }
-
-        $io->verbose('Done');
 
         return null;
     }