+++ src/Controller/Component/UploaderComponent.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace MeTools\Controller\Component;
 
 use Cake\Controller\Component;
-use Cake\Filesystem\Folder;
 use RuntimeException;
 
 /**
@@ -167,7 +166,7 @@ class UploaderComponent extends Component
         is_dir_or_fail($directory);
 
         $filename = $filename ? basename($filename) : $this->findTargetFilename($this->file->name);
-        $file = Folder::slashTerm($directory) . $filename;
+        $file = add_slash_term($directory) . $filename;
 
         if (!$this->move_uploaded_file($this->file->tmp_name, $file)) {
             $this->setError(__d('me_tools', 'The file was not successfully moved to the target directory'));