+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 2.x branch
 ## 2.18 branch
+### 2.18.9
+* added `IntegrationTestTrait::assertSessionEmpty()` method.
+
 ### 2.18.8
 * `button()`, `postButton()` and `postLink()` methods provided by `FormHelper`
     can be called with the first argument as `null`;